refactor: house all chain indexers under indexing/ with shared contract
