+++ backend/src/register-path-aliases.ts
@@ -14,8 +14,8 @@ require('module-alias').addAliases({
     '@utils': path.join(root, 'utils'),
     '@typings': path.join(root, 'types'),
     '@generated': path.join(root, 'generated'),
-    // Solana indexer lives outside src/ (a sibling under indexing/), so it
-    // resolves one level up: src/../indexing/solana → dist/src/../indexing/solana.
-    '@solana-indexer': path.join(root, '..', 'indexing', 'solana'),
+    // Chain indexers live outside src/ (a sibling under indexing/), so they
+    // resolve one level up: src/../indexing → dist/src/../indexing.
+    '@indexing': path.join(root, '..', 'indexing'),
     '@graphql': path.join(root, 'graphql', 'index.js'),
 });
